sleeping-plugin/requirements.txt
opencv-python==4.13.0.90
mediapipe==0.10.32
numpy>=2.0
pika==1.3.2
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pytest==7.4.3
pytest-asyncio==0.21.1
